Enable BFD in RR peering configs
This is non-essential, but it means that when a BFD-enabled Calico
node peers with this RR, the BFD state on that peering can go Up
instead of staying Down.

It doesn't seem to matter if the BFD state stays Down; my guess is
that it would only cause a problem if the BFD state was Up and _then_
transitioned to Down.

(Obviously it _would_ matter if we added tests to make use of the fast
BFD state detection, e.g. to quickly detect a link connectivity
breakage.  But we don't yet have any tests like that.)