Attempt to clean up as many base image files as possible
After installing on top of UBI, run a cleanup script that strips out as much as possible.  The script:

- Starts with an allow list of binaries.
- Uses ldd to find all their library dependencies.
- Removes all but a keep-list of RPMs.
- Deletes any remaining unwanted binaries.
- Deletes libraries that are not referenced by the allowed binaries.

The allow list doesn't include pkill, because it pulls in a libsystemd, which has CVEs that we want to avoid. Hence, I've switched uses of pkill over to use `sv <signal> <service>`.