AUTODETECTION_METHOD_K8S option
When running in kubernetes environment, provide the option to detect the
node addresses from the k8s node resource. This may give the most precise
autodetection option when in an environment with multiple NICs. This is
independent of the interface naming or reachability and is well suited
for runnig in policy only mode, e.g. on EKS with AWS CNI.

Precise node IP detection is necessary for eBPF to program its routes,
which then used, for instance, for connect time resolution of nodeports
both by the pods and hosts.